raxy_project/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "raxy-logic"
version = "2.0.0"
description = "Biblioteca com a lógica de negócio principal do Raxy - Clean Architecture."
readme = "README.md"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10"
dependencies = [
    "botasaurus",
    "rich",
    "python-dotenv",
    "supabase",
    "random-user-agent",
    "flask",
    "fastapi",
    "uvicorn[standard]",
    "typer[all]",
    "loguru",
    "pyyaml",
    "httpx",
    "urllib3",
    "redis",
    "requests",
    "rq",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["raxy*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
